requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.5
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
pydantic==2.4.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9 